pyproject.toml
[project]
name = "inferops-action"
version = "0.1.0"
description = "AI-powered CI/CD failure analyzer for GitHub Actions"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "openai>=1.0.0",
    "google-genai>=0.3.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]
